+++ src/pysme/solve.py
@@ -135,7 +135,7 @@ class SME_Solver:
             pass
 
     def _residuals(
-        self, param, sme, spec, uncs, mask, segments="all", isJacobian=False, **_
+        self, param, sme, spec, uncs, mask, segments="all", isJacobian=False, linelist_mode='all', **_
     ):
         """
         Calculates the synthetic spectrum with sme_func and
@@ -185,8 +185,13 @@ class SME_Solver:
         # change dynamic parameters
         if self.dynamic_param is not None:
             for name in self.dynamic_param.keys():
-                sme[name] = self.dynamic_param[name](sme)
-                print(f'Changing dynamic parameter {name} to {sme[name]:.2f}.')
+                if 'abund' in name:
+                    abund_name = name.split()[1]
+                    sme.abund[abund_name] = self.dynamic_param[name](sme) - sme.monh
+                    print(f'Changing dynamic parameter {name} to {sme.abund[abund_name]:.2f}.')
+                else:
+                    sme[name] = self.dynamic_param[name](sme)
+                    print(f'Changing dynamic parameter {name} to {sme[name]:.2f}.')
         # run spectral synthesis
         try:
             result = self.synthesizer.synthesize_spectrum(
@@ -194,9 +199,10 @@ class SME_Solver:
                 updateStructure=update,
                 reuse_wavelength_grid=reuse_wavelength_grid,
                 segments=segments,
-                passLineList=False,
+                passLineList=True,
                 updateLineList=self.update_linelist,
                 radial_velocity_mode=radial_velocity_mode,
+                linelist_mode=linelist_mode
             )
         except AtmosphereError as ae:
             # Something went wrong (left the grid? Don't go there)
@@ -250,6 +256,7 @@ class SME_Solver:
         segments="all",
         step_sizes=None,
         method="2-point",
+        linelist_mode='all',
         **_,
     ):
         """
@@ -276,7 +283,7 @@ class SME_Solver:
             abs_step=step_sizes,
             bounds=bounds,
             args=args,
-            kwargs={"isJacobian": True, "segments": segments},
+            kwargs={"isJacobian": True, "segments": segments, "linelist_mode": linelist_mode},
         )
 
         if not np.all(np.isfinite(g)):
@@ -607,7 +614,7 @@ class SME_Solver:
                 )
         return param_names
 
-    def solve(self, sme, param_names=None, segments="all", bounds=None, step_sizes=None, dynamic_param=None):
+    def solve(self, sme, param_names=None, segments="all", bounds=None, step_sizes=None, dynamic_param=None, linelist_mode='all'):
         """
         Find the least squares fit parameters to an observed spectrum
 
@@ -726,7 +733,7 @@ class SME_Solver:
                 f" ; Linelist range: {sme.linelist.wlcent.min()} - {sme.linelist.wlcent.max()} Å"
             )
 
-        # Setup LineList only once
+        # Setup LineList only once (Mingjie: not needed?)
         dll = self.synthesizer.get_dll()
         dll.SetLibraryPath()
         _ = dll.InputLineList(sme.linelist)
@@ -761,6 +768,7 @@ class SME_Solver:
                         "segments": segments,
                         "step_sizes": step_sizes,
                         "method": sme.leastsquares_jac,
+                        "linelist_mode": linelist_mode
                     },
                 )
                 # The jacobian is altered by the loss function